Add files via upload
This web-based app will first scan your current emotion with the help of media pipe and OpenCV then the proposed system will recommend movies from different ott platforms based on user behavior which is read from web cam using mediapipe, keras and opencv. this is a neural network-based approach to movie recommendation where their facial expressions detect a person's mood.  Facial expressions play a crucial role in detecting a person's mood. A webcam or camera is used to picture a face, and input is extracted from that picture. This input is also used for determining an individual's mood. a webapp is created for this using streamlit.